+++ backend/canisters/community/impl/src/updates/c2c_notify_p2p_swap_status_change.rs
@@ -1,6 +1,6 @@
 use crate::activity_notifications::handle_activity_notification;
 use crate::guards::caller_is_escrow_canister;
-use crate::{RuntimeState, execute_update};
+use crate::{CommunityEventPusher, RuntimeState, execute_update};
 use canister_api_macros::update;
 use canister_tracing_macros::trace;
 use escrow_canister::{SwapStatus, SwapStatusChange as Args};
@@ -70,6 +70,7 @@ fn c2c_notify_p2p_swap_status_change_impl(args: Args, state: &mut RuntimeState) 
                     }
                 }
                 SwapStatus::Completed(c) => {
+                    let now = state.env.now();
                     result = channel
                         .chat
                         .events
@@ -79,8 +80,12 @@ fn c2c_notify_p2p_swap_status_change_impl(args: Args, state: &mut RuntimeState) 
                             m.message_id,
                             c.token0_transfer_out.block_index,
                             c.token1_transfer_out.block_index,
-                            state.env.now(),
-                            &mut state.data.event_store_client,
+                            now,
+                            CommunityEventPusher {
+                                now,
+                                rng: state.env.rng(),
+                                queue: &mut state.data.local_user_index_event_sync_queue,
+                            },
                         )
                         .map(|result| result.drop_value())
                         .ok();