+++ backend/canisters/community/impl/src/updates/end_video_call.rs
@@ -1,7 +1,7 @@
 use crate::activity_notifications::handle_activity_notification;
 use crate::guards::caller_is_video_call_operator;
 use crate::timer_job_types::TimerJob;
-use crate::{RuntimeState, execute_update};
+use crate::{CommunityEventPusher, RuntimeState, execute_update};
 use canister_tracing_macros::trace;
 use community_canister::end_video_call_v2::*;
 use ic_cdk::update;
@@ -22,10 +22,15 @@ pub(crate) fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> OCRes
     );
 
     if let Some(channel) = state.data.channels.get_mut(&args.channel_id) {
+        let now = state.env.now();
         let result = channel.chat.events.end_video_call(
             args.message_id.into(),
-            state.env.now(),
-            Some(&mut state.data.event_store_client),
+            now,
+            Some(CommunityEventPusher {
+                now,
+                rng: state.env.rng(),
+                queue: &mut state.data.local_user_index_event_sync_queue,
+            }),
         )?;
 
         state.process_message_updated(result);