+++ backend/canisters/group/impl/src/updates/end_video_call.rs
@@ -1,7 +1,7 @@
 use crate::activity_notifications::handle_activity_notification;
 use crate::guards::caller_is_video_call_operator;
 use crate::timer_job_types::TimerJob;
-use crate::{RuntimeState, execute_update};
+use crate::{GroupEventPusher, RuntimeState, execute_update};
 use canister_tracing_macros::trace;
 use group_canister::end_video_call_v2::*;
 use ic_cdk::update;
@@ -20,10 +20,15 @@ pub(crate) fn end_video_call_impl(args: Args, state: &mut RuntimeState) -> OCRes
         },
     );
 
+    let now = state.env.now();
     state.data.chat.events.end_video_call(
         args.message_id.into(),
-        state.env.now(),
-        Some(&mut state.data.event_store_client),
+        now,
+        Some(GroupEventPusher {
+            now,
+            rng: state.env.rng(),
+            queue: &mut state.data.local_user_index_event_sync_queue,
+        }),
     )?;
 
     handle_activity_notification(state);