+++ backend/canisters/user/CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support autonomous bots without API keys ([#7985](https://github.com/open-chat-labs/open-chat/pull/7985))
 - Push notifications to LocalUserIndex canisters ([#8002](https://github.com/open-chat-labs/open-chat/pull/8002))
 - Flush pending events at the end of each update call ([#8008](https://github.com/open-chat-labs/open-chat/pull/8008))
+- Use `EventPusher` rather than `EventStoreClient` to push events ([#8010](https://github.com/open-chat-labs/open-chat/pull/8010))
 
 ## [[2.0.1747-user](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1747-user)] - 2025-05-12